missing db_free_result() calls

+++ npgCore/setup/setup-option-defaults.php
@@ -93,6 +93,7 @@ if ($rslt) {
 	while ($option = db_fetch_assoc($rslt)) {
 		query('UPDATE ' . prefix('options') . ' SET `creator`="' . str_replace('zp-core/', 'npgCore/', $option['creator']) . '"');
 	}
+	db_free_result($rslt);
 }
 
 //effervescence_plus migration
@@ -237,13 +238,16 @@ foreach ($result as $row) {
 //migrate plugin enables removing "zp" from name
 $sql = 'SELECT `id`, `name` FROM ' . prefix('options') . ' WHERE `name` LIKE "zp\_plugin\_%"';
 $result = query($sql);
-while ($row = db_fetch_assoc($result)) {
-	$sql = 'UPDATE ' . prefix('options') . ' SET `name`=' . db_quote(substr($row['name'], 2)) . ' WHERE `id`=' . $row['id'];
-	if (!query($sql, false)) {
-// the plugin has executed defaultExtension() which has set the _plugin_ option already
-		$sql = 'DELETE FROM ' . prefix('options') . ' WHERE `id`=' . $row['id'];
-		query($sql);
+if ($result) {
+	while ($row = db_fetch_assoc($result)) {
+		$sql = 'UPDATE ' . prefix('options') . ' SET `name`=' . db_quote(substr($row['name'], 2)) . ' WHERE `id`=' . $row['id'];
+		if (!query($sql, false)) {
+			// the plugin has executed defaultExtension() which has set the _plugin_ option already
+			$sql = 'DELETE FROM ' . prefix('options') . ' WHERE `id`=' . $row['id'];
+			query($sql);
+		}
 	}
+	db_free_result($result);
 }
 //clean up plugin creator field
 $sql = 'UPDATE ' . prefix('options') . ' SET `creator`=' . db_quote(CORE_FOLDER . '/setup/setup-option-defaults.php[' . __LINE__ . ']') . ' WHERE `name` LIKE "\_plugin\_%" AND `creator` IS NULL;';
@@ -266,6 +270,7 @@ if ($result) {
 			}
 		}
 	}
+	db_free_result($result);
 }
 
 //migrate "publish" dates
@@ -291,6 +296,7 @@ if ($result) {
 		$sql = 'UPDATE ' . prefix('albums') . ' SET `publishdate`=' . db_quote(date('Y-m-d H:i:s', $row['mtime'])) . ' WHERE `id`=' . $row['id'];
 		query($sql);
 	}
+	db_free_result($result);
 }
 //	fix empty sort_order
 foreach (array('news_categories', 'pages', 'images', 'albums', 'menu') as $table) {
@@ -368,6 +374,7 @@ foreach ($migrate as $file => $theme) {
 		$newcreator = str_replace($theme, $newtheme, $row['creator']);
 		query('UPDATE ' . prefix('options') . ' SET `theme`=' . db_quote($newtheme) . ', `creator`=' . db_quote($newcreator) . ' WHERE `id`=' . $row['id'], FALSE);
 	}
+	db_free_result($result);
 }
 if (in_array($_gallery->getCurrentTheme(), $migrate)) {
 	$_gallery->setCurrentTheme(substr($current_theme, 2));
